python/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.7.0
numpy>=1.24.0
Pillow>=9.5.0
matplotlib>=3.7.1
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
argparse>=1.4.0 